GH-326: [feat] Extract userId from jwt.
